Server/aws : Change server code to python
To enhance the analyze speed, change the server code from js to python.
In previous js, to run the ml python code, we should spawn the thread to
use python. It takes so much time, so change the base server code to
python. Then we don't need to spawn, just call the ml function.

Signed-off-by: 김정찬/C-Lab/삼성전자 <jc_.kim@samsung.com>